+++ R/SpaDES.config-package.R
@@ -8,6 +8,5 @@
 #' @importFrom rprojroot from_wd
 #' @importFrom rprojroot is_git_root
 #' @importFrom rprojroot is_rstudio_project
-#' @importFrom SpaDES.project getModule
 ## usethis namespace: end
 NULL